pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true

[project]
name = "omniclaw"
dynamic = ["version"]
description = "The Payment Infrastructure Layer for Autonomous AI Agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Abiola Adeshina", email = "abiolaadedayo1993@gmail.com" },
    { name = "Omnuron AI Team" }
]
keywords = [
    "ai",
    "payments",
    "usdc",
    "blockchain",
    "circle",
    "x402",
    "arc",
    "agent",
    "git",
    "mcp",
    "llm",
    "automation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "circle-developer-controlled-wallets>=9.1.0",
    "pydantic>=2.12.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.0",
    "cryptography>=44.0.0",
    "redis>=7.1.0",
    "tenacity>=8.2.0",
    # x402 official SDK and signing
    "x402[httpx]>=0.1.0",
    "eth-account>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
    "mypy>=1.19.0",
    "pre-commit>=4.2.0",
    "hatch>=1.14.1",
    "twine>=6.1.0",
    "build>=1.0.0",
    "pyinstaller>=6.3.0",
]

[project.urls]
Homepage = "https://omniclaw.ai"
Company = "https://omnuron.ai"
Documentation = "https://omniclaw.ai/docs"
Repository = "https://github.com/omnuron/omniclaw"
Issues = "https://github.com/omnuron/omniclaw/issues"

[project.scripts]
omniclaw = "omniclaw.cli:main"

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
    "uvicorn>=0.42.0",
]



[tool.hatch.build.targets.wheel]
packages = ["src/omniclaw"]

[tool.hatch.build.targets.sdist]
include = ["src/omniclaw"]

[tool.hatch.build]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
